requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
docopt==0.6.2
Flask==2.0.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==4.0.1
MarkupSafe==2.0.1
PyMySQL==1.0.2
pyrsistent==0.18.0
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
yarg==0.1.9
zipp==3.6.0
